10er

+++ ft_printf.h
@@ -12,7 +12,7 @@ int	ft_putchar(char c, int fd);
 int	ft_printf(char *str, ...);
 size_t    ft_putstr(char *s, int fd);
 int     ft_print_putnbr(long long n);
-int ft_putnbr_hex_lower(unsigned int n);
+int ft_putnbr_hex_lower(unsigned long n);
 int ft_putnbr_hex_upper(unsigned int n);
 int hexer_pointer(unsigned int n);
